requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
alembic
python-dotenv
pytest
httpx
pydantic-settings
python-multipart

